requirements.txt
django-http-proxy
beautifulsoup4